Move to formatter class
