+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/boosterinfodetails/BoosterInfoDetailsViewModel.kt
@@ -3,9 +3,7 @@ package de.rki.coronawarnapp.covidcertificate.boosterinfodetails
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
-import de.rki.coronawarnapp.ccl.dccwalletinfo.model.BoosterNotification
-import de.rki.coronawarnapp.ccl.ui.text.format
-import de.rki.coronawarnapp.ccl.ui.text.formatFaqAnchor
+import de.rki.coronawarnapp.ccl.ui.text.CCLTextFormatter
 import de.rki.coronawarnapp.covidcertificate.person.core.PersonCertificatesProvider
 import de.rki.coronawarnapp.covidcertificate.vaccination.core.repository.VaccinationRepository
 import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
@@ -20,22 +18,27 @@ class BoosterInfoDetailsViewModel @AssistedInject constructor(
     dispatcherProvider: DispatcherProvider,
     personCertificatesProvider: PersonCertificatesProvider,
     private val vaccinationRepository: VaccinationRepository,
-    @Assisted private val personIdentifierCode: String
+    @Assisted private val personIdentifierCode: String,
+    private val cclTextFormatter: CCLTextFormatter,
 ) : CWAViewModel(dispatcherProvider) {
 
     val shouldClose = SingleLiveEvent<Unit>()
 
     private val uiStateFlow = personCertificatesProvider.personCertificates.mapNotNull { certificateSet ->
-        UiState(
-            certificateSet.first { it.personIdentifier?.codeSHA256 == personIdentifierCode }
-                .dccWalletInfo!!.boosterNotification.also {
-                it.identifier?.let { id ->
-                    vaccinationRepository.acknowledgeBoosterRule(
-                        personIdentifierCode = personIdentifierCode,
-                        boosterIdentifier = id
-                    )
-                }
+        val boosterNotification = certificateSet.first { it.personIdentifier?.codeSHA256 == personIdentifierCode }
+            .dccWalletInfo!!.boosterNotification.also {
+            it.identifier?.let { id ->
+                vaccinationRepository.acknowledgeBoosterRule(
+                    personIdentifierCode = personIdentifierCode,
+                    boosterIdentifier = id
+                )
             }
+        }
+        UiState(
+            titleText = cclTextFormatter.format(boosterNotification.titleText),
+            subtitleText = cclTextFormatter.format(boosterNotification.subtitleText),
+            longText = cclTextFormatter.format(boosterNotification.longText),
+            faqUrl = cclTextFormatter.formatFaqAnchor(boosterNotification.faqAnchor)
         )
     }.catch { error ->
         // This should never happen due to checks on previous screen
@@ -45,13 +48,11 @@ class BoosterInfoDetailsViewModel @AssistedInject constructor(
     val uiState = uiStateFlow.asLiveData2()
 
     data class UiState(
-        val boosterNotification: BoosterNotification
-    ) {
-        val titleText = boosterNotification.titleText.format()
-        val subtitleText = boosterNotification.subtitleText.format()
-        val longText = boosterNotification.longText.format()
-        val faqUrl = formatFaqAnchor(boosterNotification.faqAnchor)
-    }
+        val titleText: String,
+        val subtitleText: String,
+        val longText: String,
+        val faqUrl: String?,
+    )
 
     @AssistedFactory
     interface Factory : CWAViewModelFactory<BoosterInfoDetailsViewModel> {
